+++ src/mfx/dsp/SplDataRetrievalInterface.cpp
@@ -58,7 +58,7 @@ Throws: Nothing
 ==============================================================================
 */
 
-void	SplDataRetrievalInterface::get_data (float *chn_data_ptr_arr [], int64_t pos, long len, bool invert_flag)
+void	SplDataRetrievalInterface::get_data (float *chn_data_ptr_arr [], int64_t pos, int len, bool invert_flag)
 {
 	assert (chn_data_ptr_arr != 0);
 	assert (chn_data_ptr_arr [0] != 0);
